markup: Fix summary generation mechanism
In the `for (const [title, links] of summaryMap)` loop, `links` can be
an empty array if a preceding section includes a report for the same
user. This caused a bug where a user link was ouput as `undefined`.

Also removed the mechanism that placed "Bot:" after the first section link.
It is now always generated at the beginning of the summary.