refactor(markup): use the new `leids` parameter for `list=logevents` queries
With the addition of the `leids` parameter to the `list=logevents` query
module, it is no longer necessary to scrape Special:Log to convert
log IDs to usernames.

This patch updates the IDResolver class (renamed from IDList) and
integrates logic previously handled by separate functions into it,
leveraging the `leids` parameter.